Create new package for bugsnag
