+++ server/handlers/error.go
@@ -3,9 +3,6 @@ package handlers
 import (
 	"encoding/json"
 	"net/http"
-	"os"
-
-	bugsnag "github.com/bugsnag/bugsnag-go"
 )
 
 // ErrorMessage is a JSON formatter
@@ -22,17 +19,4 @@ func HandleHTTPError(w http.ResponseWriter, errno int, err error) {
 		Code:    errno,
 		Message: err.Error(),
 	})
-}
-
-// NewBugsnag return bugsnag notifier
-func NewBugsnag() *bugsnag.Notifier {
-	errorHandlerConfig := bugsnag.Configuration{
-		APIKey:          os.Getenv("BUGSNAG_KEY"),
-		ProjectPackages: []string{"main", "github.com/Coderockr/vitrine-social/*"},
-	}
-
-	notifier := bugsnag.New(errorHandlerConfig)
-	notifier.AutoNotify()
-
-	return notifier
 }
